# contracts/openapi/piqae-v1.yaml
openapi: 3.1.0
info:
  title: Piqae API
  version: 1.0.0
  description: |
    Durable local and remote printing API.
    Every response, including framework 404/405 responses, carries `X-Request-Id`.
    A caller value is preserved only when it matches `^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$`;
    otherwise Piqae generates `req_<ULID>`. Native `error.request_id` and
    PrintNode-compatible error `uid` exactly match the response header.
    During the V1 migration window, existing `spl_*` credentials and signed
    `X-Spool-*` node headers remain accepted. New clients must use `piq_*`
    credentials and `X-Piqae-*` headers.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://api.piqae.com
    description: Piqae Cloud API origin. Self-hosted deployments replace this origin.
security:
  - bearer: []
  - platformBearer: []
...
paths:
...
  /v1/identity/local/bootstrap:
    post:
      summary: Create the first self-hosted owner
      description: |
        Available only when local-owner identity and an explicit bootstrap token are
        configured. The bootstrap token is supplied in `X-Piqae-Bootstrap-Token`.
        The returned owner credential is shown exactly once.
      tags: [Identity]
      operationId: bootstrapLocalOwner
      security: []
      parameters:
        - name: X-Piqae-Bootstrap-Token
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 512 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BootstrapLocalOwner" }
      responses:
        "201":
          description: First owner and workspace created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BootstrappedLocalOwner" }
        "401": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/identity/local/exchange:
    post:
      summary: Exchange a local-owner credential for a short-lived session
      tags: [Identity]
      operationId: exchangeLocalOwnerCredential
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExchangeLocalOwnerCredential" }
      responses:
        "200":
          description: Opaque session secret for a server-side browser adapter.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocalOwnerSession" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/identity/local/sessions/rotate:
    post:
      summary: Atomically rotate the current local-owner session
      tags: [Identity]
      operationId: rotateLocalOwnerSession
      responses:
        "200":
          description: Replacement session. The presented session is revoked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocalOwnerSession" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/identity/local/sessions/revoke:
    post:
      summary: Revoke the current local-owner session
      tags: [Identity]
      operationId: revokeLocalOwnerSession
      responses:
        "204": { description: Session revoked or already inactive. }
        "401": { $ref: "#/components/responses/Error" }
  /v1/identity/me:
    get:
      summary: Inspect the authenticated identity and active workspace
      tags: [Identity]
      operationId: getCurrentIdentity
      responses:
        "200":
          description: Tenant-scoped identity projection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentIdentity" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/workspaces/current:
    get:
      summary: Inspect the authenticated workspace
      tags: [Identity]
      operationId: getCurrentWorkspace
      responses:
        "200":
          description: Current workspace only.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/workspaces/current/members:
    get:
      summary: List members of the authenticated workspace
      tags: [Identity]
      operationId: listCurrentWorkspaceMembers
      responses:
        "200":
          description: Members scoped by the authenticated workspace ID.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkspaceMember" }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  headers:
    RequestId:
      description: Request correlation ID used by response bodies and structured completion logs.
      required: true
      schema:
        type: string
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: Piqae API key, local-owner session, or OIDC JWT
      description: API keys are scope-limited. Local-owner sessions are short-lived and self-host only. OIDC JWTs require issuer, expiry, organization, permissions, and either audience or an exact configured claim/value application binding.
    platformBearer:
      type: http
      scheme: bearer
      bearerFormat: Piqae platform service-account key
      description: |
        Operator-issued credential. Platform account-management routes use the
        bearer without tenant-selection headers. Account-scoped tenant routes
        require X-Piqae-Workspace-Id and X-Piqae-Environment-Id matching an
        active explicit grant. An ordinary API key cannot use these headers to
        change tenant.
    compatibilityBasic:
      type: http
      scheme: basic
      description: Piqae compatibility API key as the username with an empty password.
    agentId:
      type: apiKey
      in: header
      name: x-piqae-agent-id
    agentTimestamp:
      type: apiKey
      in: header
      name: x-piqae-timestamp
    agentNonce:
      type: apiKey
      in: header
      name: x-piqae-nonce
    agentBodyDigest:
      type: apiKey
      in: header
      name: x-piqae-body-sha256
    agentSignature:
      type: apiKey
      in: header
      name: x-piqae-signature
...
  responses:
...
    Error:
      description: Request failed.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    BootstrapLocalOwner:
      type: object
      additionalProperties: false
      required: [workspace_name, email]
      properties:
        workspace_name: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string, format: email, maxLength: 320 }
        display_name: { type: [string, "null"], maxLength: 120 }
    BootstrappedLocalOwner:
      type: object
      additionalProperties: false
      required: [credential, workspace, member]
      properties:
        credential:
          type: string
          pattern: '^piq_owner_'
          writeOnly: true
        workspace: { $ref: "#/components/schemas/Workspace" }
        member: { $ref: "#/components/schemas/WorkspaceMember" }
    ExchangeLocalOwnerCredential:
      type: object
      additionalProperties: false
      required: [credential]
      properties:
        credential:
          type: string
          pattern: '^piq_owner_'
          maxLength: 512
          writeOnly: true
    LocalOwnerSession:
      type: object
      additionalProperties: false
      required: [token, expires_at]
      properties:
        token:
          type: string
          pattern: '^piq_session_'
          writeOnly: true
        expires_at: { type: string, format: date-time }
    CurrentIdentity:
      type: object
      additionalProperties: false
      required: [id, email, name, workspace_id, environment_id, roles]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: [string, "null"] }
        workspace_id: { type: string }
        environment_id: { type: string }
        roles:
          type: array
          items: { type: string, enum: [owner, admin, developer, operator, viewer, billing] }
    Workspace:
      type: object
      additionalProperties: false
      required: [id, name, slug, status, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [active, suspended, cancelled] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkspaceMember:
      type: object
      additionalProperties: false
      required: [id, email, name, role, status, created_at, updated_at]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: [string, "null"] }
        role: { type: string, enum: [owner, admin, developer, operator, viewer, billing] }
        status: { type: string, enum: [pending, active, inactive] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
            retryable:
              type: boolean
            details:
              type: object